+++ p_map.c
@@ -357,12 +357,20 @@ static boolean PIT_CheckLine(line_t *ld) // killough 3/26/98: make static
   // killough 8/10/98: allow bouncing objects to pass through as missiles
   if (!(tmthing->flags & (MF_MISSILE | MF_BOUNCES)))
     {
-      if (ld->flags & ML_BLOCKING)           // explicitly blocking everything
+      // explicitly blocking everything, or blocking players (mbf21)
+      if (ld->flags & ML_BLOCKING ||
+	  (tmthing->player && ld->flags & ML_BLOCKPLAYERS))
 	return tmunstuck && !untouched(ld);  // killough 8/1/98: allow escape
 
       // killough 8/9/98: monster-blockers don't affect friends
+      // mbf21: ML_BLOCKLANDMONSTERS blocks only monsters that walk
       if (!(tmthing->flags & MF_FRIEND || tmthing->player)
-	  && ld->flags & ML_BLOCKMONSTERS)
+	  &&
+	  (
+	    ld->flags & ML_BLOCKMONSTERS ||
+	    (ld->flags & ML_BLOCKLANDMONSTERS && !(tmthing->flags & MF_FLOAT))
+	  )
+	 )
 	return false; // block monsters only
     }